+++ package-scripts.js
@@ -1,4 +1,4 @@
-const defaultConfiguration = {
+module.exports = {
     options: {
         scripts: false,
         logLevel: 'warn',
@@ -80,7 +80,7 @@ const defaultConfiguration = {
                 'concurrently --kill-others-on-fail --prefix-colors bgBlue.bold,bgMagenta.bold ' +
                 '--prefix "[{name}]" --names verdaccio.serve,publish ' +
                 '"nps verdaccio.serve" ' +
-                ('"rimraf ./test/verdaccio/storage/* ' + '&& npm publish --registry http://localhost:4567"'),
+                '"rimraf ./test/verdaccio/storage/* && npm publish --registry http://localhost:4567"',
             description: 'Run Verdaccio server and publish current version of library to it',
             serve: {
                 script: 'verdaccio --config ./test/verdaccio/config.yml',
@@ -98,8 +98,24 @@ const defaultConfiguration = {
                 description: 'Generate changelog based on tags, starting from scratch',
                 hiddenFromHelp: true
             }
+        },
+
+        license: {
+            script:
+                'npx tsconfig.js --once --root=./project-types/common/license.config.js --add-comments=none ' +
+                '&& license-check-and-add add -f ./project-types/common/license.config.json ' +
+                '&& npx rimraf ./project-types/common/license.config.json',
+            hiddenFromHelp: true,
+            description: 'Add license information to all code files in the project',
+            remove: {
+                script:
+                    'npx tsconfig.js --once --root=./project-types/common/license.config.js --add-comments=none ' +
+                    '&& license-check-and-add remove -f ./project-types/common/license.config.json ' +
+                    '&& npx rimraf ./project-types/common/license.config.json',
+
+                hiddenFromHelp: true,
+                description: 'Add license information to all code files in the project'
+            }
         }
     }
 };
-
-module.exports = defaultConfiguration;