+++ project-types/webLibrary/demos/list.js
@@ -1,7 +1,0 @@
-demo_list = [
-    {
-        index: 'basic_demo/index.html',
-        title: 'Simple demo',
-        description: 'A simple demo for this webpage'
-    }
-];